+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Maintenance mode Changelog
 
+## 1.0.2 - 2020-01-14
+- Added ACP settings and ACP module.
+- Added social link for contact.
+
 ## 1.0.1 - 2019-12-21
 - Added login button
-- Twig update.
+- Twig update
 
 ## 1.0.0 - 2019-12-20
 - First release